pyproject.toml
[tool.poetry]
name = "get-my-id-bot"
version = "0.1.0"
description = "Telegram bot for getting user and chat IDs"
authors = ["Daniil Pavlovich <layred.dota2@mail.ru>"]
readme = "README.md"
packages = [{include = "bot", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.7"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
